Renamed module consensus-event-creator, setup code ownership of new consensus modules
