+++ platform-sdk/consensus-event-creator-impl/build.gradle.kts
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: Apache-2.0
+plugins {
+    id("org.hiero.gradle.module.library")
+    id("org.hiero.gradle.feature.publish-artifactregistry")
+}
+
+mainModuleInfo { annotationProcessor("com.swirlds.config.processor") }
+
+description = "Default Consensus Event Creator Implementation"
