+++ platform-sdk/consensus-event-creator-impl/src/main/java/org/hiero/consensus/event/creator/impl/EventCreatorImpl.java
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: Apache-2.0
+package org.hiero.consensus.event.creator.impl;
+
+import org.hiero.consensus.event.creator.EventCreator;
+
+/**
+ * Implementation for the mock event creator.
+ */
+public class EventCreatorImpl implements EventCreator {}
