+++ platform-sdk/consensus-event-creator/build.gradle.kts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+plugins {
+    id("org.hiero.gradle.module.library")
+    id("org.hiero.gradle.feature.publish-artifactregistry")
+}
+
+description = "Consensus Event Creator API"
