+++ platform-sdk/consensus-event-creator/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+module org.hiero.consensus.event.creator {
+    exports org.hiero.consensus.event.creator;
+
+    requires static transitive com.github.spotbugs.annotations;
+}
