+++ platform-sdk/consensus-event-creator/src/main/java/org/hiero/consensus/event/creator/EventCreator.java
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+package org.hiero.consensus.event.creator;
+
+/**
+ * Mock event creator.
+ */
+public interface EventCreator {}
