--- platform-sdk/event-creator-impl/build.gradle.kts
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-plugins {
-    id("org.hiero.gradle.module.library")
-    id("org.hiero.gradle.feature.publish-artifactregistry")
-}
-
-mainModuleInfo { annotationProcessor("com.swirlds.config.processor") }
-
-description = "Default Event Creator Implementation"
