--- platform-sdk/event-creator-impl/src/main/java/org/hiero/event/creator/impl/EventCreatorImpl.java
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-package org.hiero.event.creator.impl;
-
-import org.hiero.event.creator.EventCreator;
-
-/**
- * Implementation for the mock event creator.
- */
-public class EventCreatorImpl implements EventCreator {}
