--- platform-sdk/event-creator/build.gradle.kts
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-plugins {
-    id("org.hiero.gradle.module.library")
-    id("org.hiero.gradle.feature.publish-artifactregistry")
-}
-
-description = "Event Creator API"
