--- platform-sdk/event-creator/src/main/java/module-info.java
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-module org.hiero.event.creator {
-    exports org.hiero.event.creator;
-
-    requires static transitive com.github.spotbugs.annotations;
-}
