--- platform-sdk/event-creator/src/main/java/org/hiero/event/creator/EventCreator.java
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-package org.hiero.event.creator;
-
-/**
- * Mock event creator.
- */
-public interface EventCreator {}
