+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/config/PlatformConfigurationExtension.java
@@ -32,8 +32,8 @@ import com.swirlds.platform.wiring.PlatformSchedulersConfig;
 import com.swirlds.virtualmap.config.VirtualMapConfig;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Set;
+import org.hiero.consensus.event.creator.impl.EventCreationConfig;
 import org.hiero.consensus.model.node.NodeId;
-import org.hiero.event.creator.impl.EventCreationConfig;
 
 /**
  * Registers configuration types for the platform.
