+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/creation/DefaultEventCreationManager.java
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.swirlds.platform.event.creation;
 
-import static org.hiero.event.creator.EventCreationStatus.ATTEMPTING_CREATION;
-import static org.hiero.event.creator.EventCreationStatus.IDLE;
-import static org.hiero.event.creator.EventCreationStatus.NO_ELIGIBLE_PARENTS;
-import static org.hiero.event.creator.EventCreationStatus.RATE_LIMITED;
+import static org.hiero.consensus.event.creator.EventCreationStatus.ATTEMPTING_CREATION;
+import static org.hiero.consensus.event.creator.EventCreationStatus.IDLE;
+import static org.hiero.consensus.event.creator.EventCreationStatus.NO_ELIGIBLE_PARENTS;
+import static org.hiero.consensus.event.creator.EventCreationStatus.RATE_LIMITED;
 
 import com.swirlds.common.context.PlatformContext;
 import com.swirlds.common.metrics.extensions.PhaseTimer;
@@ -17,16 +17,16 @@ import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import org.hiero.consensus.event.creator.EventCreationRule;
+import org.hiero.consensus.event.creator.EventCreationStatus;
+import org.hiero.consensus.event.creator.impl.EventCreationConfig;
+import org.hiero.consensus.event.creator.impl.rules.AggregateEventCreationRules;
+import org.hiero.consensus.event.creator.impl.rules.MaximumRateRule;
+import org.hiero.consensus.event.creator.impl.rules.PlatformHealthRule;
 import org.hiero.consensus.model.event.PlatformEvent;
 import org.hiero.consensus.model.event.UnsignedEvent;
 import org.hiero.consensus.model.hashgraph.EventWindow;
 import org.hiero.consensus.model.status.PlatformStatus;
-import org.hiero.event.creator.EventCreationRule;
-import org.hiero.event.creator.EventCreationStatus;
-import org.hiero.event.creator.impl.EventCreationConfig;
-import org.hiero.event.creator.impl.rules.AggregateEventCreationRules;
-import org.hiero.event.creator.impl.rules.MaximumRateRule;
-import org.hiero.event.creator.impl.rules.PlatformHealthRule;
 
 /**
  * Default implementation of the {@link EventCreationManager}.
