+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/creation/rules/PlatformStatusRule.java
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.swirlds.platform.event.creation.rules;
 
-import static org.hiero.event.creator.EventCreationStatus.PLATFORM_STATUS;
+import static org.hiero.consensus.event.creator.EventCreationStatus.PLATFORM_STATUS;
 
 import com.swirlds.platform.pool.TransactionPoolNexus;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Objects;
 import java.util.function.Supplier;
+import org.hiero.consensus.event.creator.EventCreationRule;
+import org.hiero.consensus.event.creator.EventCreationStatus;
 import org.hiero.consensus.model.status.PlatformStatus;
-import org.hiero.event.creator.EventCreationRule;
-import org.hiero.event.creator.EventCreationStatus;
 
 /**
  * Limits the creation of new events depending on the current platform status.
