+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/pool/TransactionPoolNexus.java
@@ -20,8 +20,8 @@ import java.util.Objects;
 import java.util.Queue;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.hiero.consensus.event.creator.impl.EventCreationConfig;
 import org.hiero.consensus.model.status.PlatformStatus;
-import org.hiero.event.creator.impl.EventCreationConfig;
 
 /**
  * Store a list of transactions created by self, both system and non-system, for wrapping in the next event to be
