+++ platform-sdk/swirlds-platform-core/src/main/java/module-info.java
@@ -139,9 +139,9 @@ module com.swirlds.platform.core {
     requires transitive com.swirlds.config.api;
     requires transitive com.swirlds.metrics.api;
     requires transitive com.swirlds.state.api;
+    requires transitive org.hiero.consensus.event.creator;
     requires transitive org.hiero.consensus.gossip;
     requires transitive org.hiero.consensus.model;
-    requires transitive org.hiero.event.creator;
     requires transitive com.fasterxml.jackson.annotation;
     requires transitive com.fasterxml.jackson.databind;
     requires transitive info.picocli;
@@ -151,8 +151,8 @@ module com.swirlds.platform.core {
     requires com.swirlds.merkle;
     requires com.swirlds.merkledb;
     requires com.swirlds.virtualmap;
+    requires org.hiero.consensus.event.creator.impl;
     requires org.hiero.consensus.utility;
-    requires org.hiero.event.creator.impl;
     requires com.fasterxml.jackson.core;
     requires com.fasterxml.jackson.dataformat.yaml;
     requires com.github.spotbugs.annotations;
