+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/pool/TransactionPoolNexusTest.java
@@ -16,8 +16,8 @@ import com.swirlds.config.api.Configuration;
 import com.swirlds.platform.config.TransactionConfig;
 import java.time.Duration;
 import java.util.List;
+import org.hiero.consensus.event.creator.impl.EventCreationConfig;
 import org.hiero.consensus.model.status.PlatformStatus;
-import org.hiero.event.creator.impl.EventCreationConfig;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
